# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "第一个gin项目",
        "title": "BlueBell",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8081",
    "basePath": "/api/v1",
    "paths": {
        "/community": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取所有社区",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "社区相关"
                ],
                "summary": "获取所有社区",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers._ResponseCommunityList"
                        }
                    }
                }
            }
        },
        "/community/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "社区详情接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "社区相关"
                ],
                "summary": "社区详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "社区id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers._ResponseCommunityDetail"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "用户登陆接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "登陆",
                "parameters": [
                    {
                        "description": "登陆参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ParamLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登陆成功",
                        "schema": {
                            "$ref": "#/definitions/controllers._ResponseLogin"
                        }
                    }
                }
            }
        },
        "/posts2": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "可按社区按时间或分数排序查询帖子列表接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帖子相关"
                ],
                "summary": "升级版帖子列表接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "社区ID  必填",
                        "name": "community_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "time",
                        "description": "排序    选填 默认:time",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "条数    选填 默认:10",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers._ResponsePostList"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "用户注册接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ParamSignUp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功数据",
                        "schema": {
                            "$ref": "#/definitions/controllers._ResponseCodeWithMsg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers._ResponseCodeWithMsg": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务响应状态码",
                    "type": "integer"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
        "controllers._ResponseCommunityDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "object",
                    "$ref": "#/definitions/model.CommunityDetail"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
        "controllers._ResponseCommunityList": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Community"
                    }
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
        "controllers._ResponseLogin": {
            "type": "object",
            "properties": {
                "AToken": {
                    "type": "string"
                },
                "RToken": {
                    "type": "string"
                },
                "code": {
                    "description": "业务响应状态码",
                    "type": "integer"
                },
                "msg": {
                    "description": "提示信息",
                    "type": "string"
                },
                "use_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "controllers._ResponsePostList": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ApiPostDetail"
                    }
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
        "model.ApiPostDetail": {
            "type": "object",
            "required": [
                "community_id",
                "content"
            ],
            "properties": {
                "author_id": {
                    "description": "作者id",
                    "type": "integer"
                },
                "author_name": {
                    "type": "string"
                },
                "community_id": {
                    "description": "社区id",
                    "type": "integer"
                },
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "introduction": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "description": "帖子状态",
                    "type": "integer"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "vote_num": {
                    "type": "integer"
                }
            }
        },
        "model.Community": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.CommunityDetail": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "introduction": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.ParamLogin": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.ParamPostList": {
            "type": "object",
            "properties": {
                "community_id": {
                    "description": "社区ID  必填",
                    "type": "integer"
                },
                "page": {
                    "description": "排序    选填 默认:time",
                    "type": "string",
                    "example": "time"
                },
                "size": {
                    "description": "条数    选填 默认:10",
                    "type": "integer"
                }
            }
        },
        "model.ParamSignUp": {
            "type": "object",
            "required": [
                "password",
                "repassword",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码   必填",
                    "type": "string"
                },
                "repassword": {
                    "description": "确认密码 必填",
                    "type": "string"
                },
                "username": {
                    "description": "用户名 必填",
                    "type": "string"
                }
            }
        }
    }
}